requirements.txt


python-dotenv==1.1.0
psycopg2-binary==2.9.10
django==5.2.3
djangorestframework==3.16.0
requests==2.32.4
fake-useragent==2.2.0